refactor: blog content section을 기능별 컴포넌트로 분리 (title, tabNavigation, rulerDecoration)
