+++ public/sitemap-0.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
-<url><loc>https://www.depromeet.com</loc><lastmod>2025-05-27T17:55:38.724Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.depromeet.com/about</loc><lastmod>2025-05-27T17:55:38.725Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.depromeet.com/apply</loc><lastmod>2025-05-27T17:55:38.725Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.depromeet.com/blog</loc><lastmod>2025-05-27T17:55:38.725Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.depromeet.com/project</loc><lastmod>2025-05-27T17:55:38.725Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.depromeet.com/recruit</loc><lastmod>2025-05-27T17:55:38.725Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.depromeet.com</loc><lastmod>2025-06-15T09:02:37.447Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.depromeet.com/about</loc><lastmod>2025-06-15T09:02:37.448Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.depromeet.com/apply</loc><lastmod>2025-06-15T09:02:37.448Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.depromeet.com/blog</loc><lastmod>2025-06-15T09:02:37.448Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.depromeet.com/project</loc><lastmod>2025-06-15T09:02:37.448Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.depromeet.com/recruit</loc><lastmod>2025-06-15T09:02:37.448Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
 </urlset>
