+++ src/features/Blog/sections/BlogContentSection.tsx
@@ -1,13 +1,14 @@
-import { useEffect, useRef, useState } from 'react';
-import Image from 'next/image';
+import { useEffect, useState } from 'react';
 import { css } from '@emotion/react';
 
 import { AllBlog, DEEPER_BLOG_LIST, OFFICIAL_BLOG_LIST } from '~/constant/blog';
 import { sectionGridBg } from '~/styles/background';
 import { mediaQuery } from '~/styles/media';
-import { theme } from '~/styles/theme';
 
 import { BlogPaginationSection } from './BlogPaginationSection';
+import { BlogRulerDecoration } from './BlogRulerDecoration';
+import { BlogTabNavigation } from './BlogTabNavigation';
+import { BlogTitleSection } from './BlogTitleSection';
 
 const OFFICIAL_SUB_TABS = [
   { key: 'entire', name: '전체' },
@@ -39,7 +40,6 @@ export const BlogContentSection = () => {
   const [currentMainTab, setCurrentMainTab] = useState<MainTabsKeys>('OFFICIAL');
   const mainTab = mainTabs[currentMainTab];
   const [currentSubTab, setCurrentSubTab] = useState(mainTab.subTabs[0]);
-  const activeTabRefs = useRef<{ [key: string]: HTMLButtonElement | null }>({});
 
   const getTabKeyByName = (name: string): MainTabsKeys => {
     return (
@@ -58,28 +58,6 @@ export const BlogContentSection = () => {
     setCurrentSubTab(matched ?? mainTab.subTabs[0]);
   };
 
-  // 언더라인 길이 조정
-  useEffect(() => {
-    const activeButton = activeTabRefs.current[currentSubTab.name];
-    if (activeButton) {
-      // 텍스트 길이 측정을 위한 임시 span 생성
-      const span = document.createElement('span');
-      span.textContent = currentSubTab.name;
-      span.style.visibility = 'hidden';
-      span.style.position = 'absolute';
-      span.style.fontSize = '16px';
-      span.style.fontFamily = 'Pretendard';
-      span.style.fontWeight = '600';
-      document.body.appendChild(span);
-
-      const textWidth = span.offsetWidth;
-      document.body.removeChild(span);
-
-      // CSS 변수로 텍스트 너비 설정
-      activeButton.style.setProperty('--text-width', `${textWidth}px`);
-    }
-  }, [currentSubTab]);
-
   useEffect(() => {
     setCurrentSubTab(mainTab.subTabs[0]);
   }, [mainTab.subTabs]);
@@ -92,59 +70,22 @@ export const BlogContentSection = () => {
 
   return (
     <section css={sectionCss}>
-      <div css={titleImageContainerCss}>
-        <img
-          src="/images/blog/17th-blog.png"
-          alt="Blog"
-          width={190}
-          height={84}
-          css={titleImageCss}
-        />
-      </div>
-      <div css={tabContainerCss}>
-        {/* 메인 탭 */}
-        <div css={mainTabWrapperCss}>
-          {Object.values(mainTabs).map(({ name }) => (
-            <button
-              key={name}
-              css={[mainTabItemCss, mainTab.name === name && mainTabItemActiveCss]}
-              onClick={() => handleClickMainTab(name)}
-            >
-              {name}
-            </button>
-          ))}
-        </div>
+      <BlogTitleSection />
 
-        {/* 서브 탭 */}
-        <div css={subTabWrapperCss}>
-          {mainTab.subTabs.map(({ name }) => (
-            <button
-              key={name}
-              ref={el => {
-                activeTabRefs.current[name] = el;
-              }}
-              css={[subTabItemCss, currentSubTab.name === name && subTabItemActiveCss]}
-              onClick={() => handleClickSubTab(name)}
-            >
-              {name}
-            </button>
-          ))}
-        </div>
-      </div>
+      <BlogTabNavigation
+        currentMainTab={mainTab.name}
+        currentSubTab={currentSubTab}
+        mainTabs={mainTabs}
+        onMainTabClick={handleClickMainTab}
+        onSubTabClick={handleClickSubTab}
+      />
 
       <BlogPaginationSection
         key={`${currentMainTab}-${currentSubTab.key}-${filteredBlogList.length}`}
         blogList={filteredBlogList}
       />
 
-      <div css={rulerImageContainerCss}>
-        <Image
-          src="/images/blog/footer-ruler.png"
-          alt="ruler decoration"
-          fill
-          css={rulerImageCss}
-        />
-      </div>
+      <BlogRulerDecoration />
     </section>
   );
 };
@@ -166,184 +107,4 @@ const sectionCss = css`
     padding: 112px 20px;
     gap: 32px;
   }
-`;
-
-const titleImageContainerCss = css`
-  align-self: flex-start;
-  max-width: 1200px;
-  width: 100%;
-  padding-inline: 48px;
-  padding-top: 47px;
-  padding-bottom: 26px;
-
-  ${mediaQuery('mobile')} {
-    padding-left: 20px;
-  }
-`;
-
-const titleImageCss = css`
-  object-fit: contain;
-
-  ${mediaQuery('mobile')} {
-    width: 150px;
-    height: 60px;
-  }
-`;
-
-const tabContainerCss = css`
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-  width: 100%;
-  max-width: 1200px;
-  margin: 0 auto;
-  align-items: center;
-`;
-
-const mainTabWrapperCss = css`
-  display: flex;
-  width: 578px;
-  border-radius: 0;
-  overflow: hidden;
-  border: 1px solid #478af4;
-
-  ${mediaQuery('tablet')} {
-    width: 100%;
-    max-width: 500px;
-  }
-  ${mediaQuery('mobile')} {
-    flex-direction: column;
-    width: 100%;
-    max-width: 400px;
-  }
-`;
-
-const mainTabItemCss = css`
-  flex: 1;
-  padding: 16px 24px;
-  background-color: #e5e7eb;
-  color: black;
-  border: none;
-  cursor: pointer;
-  transition: all 0.2s ease;
-  ${theme.typosV2.pretendard.medium16};
-
-  &:not(:last-child) {
-    border-right: 1px solid #d1d5db;
-  }
-
-  &:hover {
-    background-color: #d1d5db;
-  }
-
-  ${mediaQuery('mobile')} {
-    padding: 14px 20px;
-    ${theme.typosV2.pretendard.medium14};
-
-    &:not(:last-child) {
-      border-right: none;
-      border-bottom: 1px solid #d1d5db;
-    }
-  }
-`;
-
-const mainTabItemActiveCss = css`
-  background-color: #478af4;
-  color: white;
-
-  &:hover {
-    background-color: #2563eb;
-  }
-`;
-
-const subTabWrapperCss = css`
-  display: flex;
-  width: 578px;
-  background-color: #e3e5ea;
-  padding: 0 48px;
-  justify-content: center;
-  gap: 10px;
-
-  ${mediaQuery('tablet')} {
-    width: 100%;
-    max-width: 500px;
-    padding: 0 24px;
-  }
-  ${mediaQuery('mobile')} {
-    padding: 0 12px;
-    overflow-x: auto;
-    width: 100%;
-    max-width: 400px;
-
-    &::-webkit-scrollbar {
-      display: none;
-    }
-  }
-`;
-
-const subTabItemCss = css`
-  padding: 16px 20px;
-  background: #e3e5ea;
-  color: #9595a1;
-  border: none;
-  cursor: pointer;
-  position: relative;
-  transition: all 0.2s ease;
-  ${theme.typosV2.pretendard.semibold16};
-  white-space: nowrap;
-  flex: 1 0 0;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-  &:hover {
-    color: black;
-  }
-
-  ${mediaQuery('mobile')} {
-    padding: 14px 12px;
-    ${theme.typosV2.pretendard.medium13};
-    flex-shrink: 0;
-  }
-`;
-
-const subTabItemActiveCss = css`
-  color: ${theme.colors.grey['900']};
-  font-weight: 600;
-
-  &::after {
-    content: '';
-    position: absolute;
-    bottom: 12px;
-    left: 50%;
-    transform: translateX(-50%);
-    width: var(--text-width, 40px);
-    height: 2px;
-    background-color: ${theme.colors.grey['900']};
-    transition: width 0.2s ease;
-
-    ${mediaQuery('mobile')} {
-      bottom: 12px;
-    }
-  }
-`;
-
-const rulerImageContainerCss = css`
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  width: 100%;
-  height: 20px;
-  z-index: 1;
-
-  ${mediaQuery('mobile')} {
-    height: 15px;
-  }
-`;
-
-const rulerImageCss = css`
-  object-fit: cover;
-  object-position: center;
 `;
